uff, `from` is a keyword of course
and there is where the Python API gets ugly. so we could rename it to
`source` and `target` like in the openLCA API. but it should be `from` and `to` in
the JSON serialization. so maybe we introduce a `proto` field in olca-schema to
support field renaming? `from` and `to` are not so good names... we should rename
it later...